Migrate share helper to TS
